pyproject.toml
[tool.poetry]
name = "virtual-bookshelf"
version = "0.1.0"
description = "A virtual bookshelf."
authors = ["Igor Ferreira <iferreira.n01@gmail.com>"]
readme = "README.md"
packages = [{include = "virtual_bookshelf"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
bootstrap-flask = "^2.2.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.6"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
mypy = "^1.1.1"
taskipy = "^1.10.3"
pytest = "^7.3.1"
coverage = "^7.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "isort --check --diff . && blue --check --diff ."
format = "isort . && blue ."
pre_test = "task lint"
test = "coverage run --source=virtual_bookshelf -m pytest -s -x -vv"
post_test = "coverage html"

[tool.mypy]
packages = [
    'virtual_bookshelf'
]

[[tool.mypy.overrides]]
module = "flask_bootstrap.*"
ignore_missing_imports = true